+++ source/client/inc/clientInt.h
@@ -59,7 +59,7 @@ typedef struct SAppInstInfo {
   SCorEpSet         mgmtEp;
   SInstanceActivity summary;
   SList            *pConnList;  // STscObj linked list
-  uint32_t          clusterId;
+  int64_t           clusterId;
   void             *pTransporter;
 } SAppInstInfo;
